agents/security_audit_agent/requirements.txt
smolagents
smolagents[openai]
smolagents[litellm]
ollama
python-dotenv
google-genai